fix: 修正 QuizService 傳入 long 型別給 int 參數導致編譯失敗
findRandomWordsFromOtherBooks 的 limit 參數為 int，
移除多餘的 (long) 強制轉型。

+++ src/main/java/com/linguamastery/api/service/QuizService.java
@@ -72,7 +72,7 @@ public class QuizService {
         // 不足 3 個時從其他書補
         if (wrongOptions.size() < 3) {
             int needed = 3 - wrongOptions.size();
-            wordRepository.findRandomWordsFromOtherBooks(userId, bookId, (long) needed)
+            wordRepository.findRandomWordsFromOtherBooks(userId, bookId, needed)
                     .stream().map(Word::getTranslation).forEach(wrongOptions::add);
         }
